+++ functions/registration.go
@@ -7,9 +7,9 @@ package functions
 import (
 	"context"
 
-	"hmruntime/host"
 	"hmruntime/logger"
 	"hmruntime/plugins"
+	"hmruntime/wasmhost"
 )
 
 var Functions = make(map[string]FunctionInfo)
@@ -23,7 +23,7 @@ func MonitorRegistration(ctx context.Context) {
 	go func() {
 		for {
 			select {
-			case <-host.RegistrationRequest:
+			case <-wasmhost.RegistrationRequest:
 				r := newRegistration()
 				r.registerAll(ctx)
 				r.cleanup(ctx)
@@ -47,7 +47,7 @@ func newRegistration() *registration {
 }
 
 func (r *registration) registerAll(ctx context.Context) {
-	var plugins = host.Plugins.GetAll()
+	var plugins = wasmhost.Plugins.GetAll()
 	for _, plugin := range plugins {
 		r.registerPlugin(ctx, &plugin)
 	}
