+++ graphql/graphql.go
@@ -9,15 +9,15 @@ import (
 	"net/http"
 
 	"hmruntime/graphql/engine"
-	"hmruntime/host"
 	"hmruntime/logger"
+	"hmruntime/wasmhost"
 
 	gql "github.com/wundergraph/graphql-go-tools/execution/graphql"
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/graphqlerrors"
 )
 
 func Initialize() {
-	host.RegisterPluginLoadedCallback(engine.Activate)
+	wasmhost.RegisterPluginLoadedCallback(engine.Activate)
 }
 
 func HandleGraphQLRequest(w http.ResponseWriter, r *http.Request) {
