+++ main.go
@@ -14,10 +14,10 @@ import (
 	"hmruntime/config"
 	"hmruntime/functions"
 	"hmruntime/graphql"
-	"hmruntime/host"
 	"hmruntime/logger"
 	"hmruntime/server"
 	"hmruntime/storage"
+	"hmruntime/wasmhost"
 
 	"github.com/joho/godotenv"
 )
@@ -45,14 +45,14 @@ func main() {
 	}
 
 	// Initialize the WebAssembly runtime
-	err = host.InitWasmRuntime(ctx)
+	err = wasmhost.InitWasmRuntime(ctx)
 	if err != nil {
 		log.Fatal().Err(err).Msg("Failed to initialize the WebAssembly runtime.  Exiting.")
 	}
-	defer host.WasmRuntime.Close(ctx)
+	defer wasmhost.RuntimeInstance.Close(ctx)
 
 	// Connect Hypermode host functions
-	err = functions.InstantiateHostFunctions(ctx, host.WasmRuntime)
+	err = functions.InstantiateHostFunctions(ctx, wasmhost.RuntimeInstance)
 	if err != nil {
 		log.Fatal().Err(err).Msg("Failed to instantiate host functions.  Exiting.")
 	}
@@ -67,7 +67,7 @@ func main() {
 	appdata.MonitorAppDataFiles(ctx)
 
 	// Load plugins and monitor for changes
-	host.MonitorPlugins(ctx)
+	wasmhost.MonitorPlugins(ctx)
 
 	// Initialize the GraphQL engine
 	graphql.Initialize()
